Step R42: Updates moved to Tier B, and originals now cover unique indexes (#56)
The class re-parents onto UpdatesRelationalTestBase and the fixture onto its nested
UpdatesRelationalFixture. 28 tests become 36, all green. Full local run Passed: 27058,
Failed: 82, Total: 27375 -- FIXED none, BROKEN none by name. failed unchanged at 82, total
27367 -> 27375. Compliance missing bases 46 -> 45.

Three Tier A overrides are deleted because the move makes them false: both concurrency messages
(they were InMemoryStrings', and the relational base states them itself), the reseed override
whose remark said the InMemory store "has no transaction to roll back", and EF issue #29875's,
which was InMemory's alone. UseTransaction is written in the same commit, per D6.

The move found a second defect of R40's family, fixed here. Swap_filtered_unique_index_values and
Swap_computed_unique_index_values swap the values of a unique index between two rows, and the
store answered SQLite Error 19: 'UNIQUE constraint failed: Products.Name, Products.Price'.
CommandBatchPreparer orders by value dependencies as well as row ones -- one row must release a
unique value before another may take it -- and the only thing that says which row is releasing
what is the original. ChangeEntryMapper sent originals for concurrency tokens and foreign keys;
neither Name nor Price is either. The condition now also admits a property contained in a unique
index, and both tests pass.

Not convergence, checked rather than assumed: UpdatesSqliteTest overrides only
Save_with_shared_foreign_key and Identifiers_are_generated_correctly, so EF's own SQLite run
passes both swap tests.

One expectation was wrong and the measurement corrected it.
Identifiers_are_generated_correctly asserts GetTableName() on the client's model and was
predicted to hit the M9 boundary. It passes: EF keeps the table name as a core annotation, so a
client that builds no relational model still has it.

Also records R43: Translations is priced and deliberately not moved. EF's SQLite Translations
suite carries 217 public overrides across six classes, nearly all of them SQLite lacking a
function, which says something about the store and nothing about this provider. The Tier A family
is 333 tests green today and exercises what is this provider's concern. Left for the owner,
priced rather than argued.

src/, so both gates ran: trim ratchet OK (89 <= 89, unchanged) and CI=true Release build clean.
